# src/main/resources/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Newsroom service
  version: "1.0"
  description: Newsroom API
tags:
  - name: News
    description: News management

servers:
  - url: http://localhost:8080
    description: Local server


paths:
  /newsroom-service/api/v1/summaries:
    post:
      tags:
        - Summary
      operationId: generateArticleSummary
      description: Generates a summary for a given article, initial language could be any language but the output language should one of the supported languages or EN by default
      requestBody:
        $ref: "#/components/requestBodies/GenerateSummaryRequest"
      responses:
        "201":
          description: Returns generated article summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      parameters:
        - $ref: "#/components/parameters/PageQueryParameter"
        - $ref: "#/components/parameters/SizeQueryParameter"
      tags:
        - Summary
      operationId: retrieveSummaries
      responses:
        "200":
          description: Returns all summaries with pagination
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Summary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /summaries/{summariesId}:
    parameters:
      - name: summariesId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Summary
      operationId: retrieveSummary
      responses:
        "200":
          description: Returns summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    PageQueryParameter:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeQueryParameter:
      name: size
      in: query
      schema:
        type: integer
        minimum: 10
        maximum: 30
        default: 10
  # ========================================
  # ===== Request bodies
  # ========================================
  requestBodies:
    GenerateSummaryRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NewSummary"
      required: true

  # ========================================
  # ===== Schemas
  # ========================================
  schemas:
    NewSummary:
      type: object
      properties:
        article:
          type: string
          description: Article to summarize
        outputLanguage:
          $ref: "#/components/schemas/Language"
          default: EN
          description: Language which will be used to generate the summary, default is EN
      required:
        - article
        - outputLanguage
    Summary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        language:
          $ref: "#/components/schemas/Language"
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        deletedAt:
          type: string
          format: date-time
          readOnly: true
      required:
        - id
        - title
        - content
        - tags
        - language
        - createdAt
    Language:
      enum:
        - EN
        - PT
      default: EN
    Error:
      required:
        - message
      properties:
        message:
          type: string
        errorCode:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'

    FieldError:
      required:
        - message
        - field
      properties:
        errorCode:
          type: string
        message:
          type: string
        field:
          type: string

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnsupportedMediaType:
      description: Unsupported Media Type
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotAcceptable:
      description: Not Acceptable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"